[Chore/#76] SignupStep enum 분리
